requirements.txt
# Agent Gym - Dependencies
# Minimal set for 12-hour hackathon

# Gemini SDK
google-generativeai>=0.3.0

# Data validation
pydantic>=2.0.0

# Utilities
numpy>=1.24.0

# Pretty console output
rich>=13.0.0

